refactor: Using black formatter to PEP-8 structure
