+++ main.py
@@ -1,0 +1,7 @@
+"""Main file for the application."""
+
+from src.app import App
+
+if __name__ == "__main__":
+    app = App()
+    app.mainloop()
